requirements.txt
dlib==19.16.0
numpy==1.16.1
Pillow==5.4.1
